Add method to get account's featured tags
