+++ entities/src/tag.rs
@@ -1,6 +1,7 @@
-use crate::conversion;
+use crate::{conversion, TagId};
 use serde::{Deserialize, Serialize};
-use time::Date;
+use time::{serde::iso8601, Date, OffsetDateTime};
+use url::Url;
 
 /// Represents a hashtag used within the content of a status.
 ///
@@ -33,10 +34,28 @@ pub struct History {
     pub accounts: u64,
 }
 
+/// Represents a hashtag that is featured on a profile.
+///
+/// See also [the API documentation](https://docs.joinmastodon.org/entities/FeaturedTag/)
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Featured {
+    /// The internal ID of the featured tag in the database.
+    pub id: TagId,
+    /// The name of the hashtag being featured.
+    pub name: String,
+    /// A link to all statuses by a user that contain this hashtag.
+    pub url: Url,
+    /// The number of authored statuses containing this hashtag.
+    pub statuses_count: u64,
+    /// The timestamp of the last authored status containing this hashtag.
+    #[serde(with = "iso8601")]
+    pub last_status_at: OffsetDateTime,
+}
+
 #[cfg(test)]
 mod tests {
-    use time::Month;
-
+    use time::{Month, format_description::well_known::Iso8601};
+    
     use super::*;
 
     #[test]
@@ -93,5 +112,29 @@ mod tests {
         assert_eq!(entry.uses, 200);
         assert_eq!(entry.accounts, 31);
         assert_eq!(subject.following, Some(false));
+    }
+
+    #[test]
+    fn test_featured_tag() {
+        let example = r#"{
+            "id": "627",
+            "name": "nowplaying",
+            "url": "https://mastodon.social/@trwnh/tagged/nowplaying",
+            "statuses_count": 70,
+            "last_status_at": "2022-08-29T12:03:35.061Z"
+        }"#;
+        let subject: super::Featured = serde_json::from_str(example).expect("deserialize");
+        assert_eq!(subject.id, TagId::new("627"));
+        assert_eq!(subject.name, "nowplaying");
+        assert_eq!(
+            subject.url.as_ref(),
+            "https://mastodon.social/@trwnh/tagged/nowplaying"
+        );
+        assert_eq!(subject.statuses_count, 70);
+        assert_eq!(
+            subject.last_status_at,
+            OffsetDateTime::parse("2022-08-29T12:03:35.061Z", &Iso8601::PARSING)
+                .expect("parse test time")
+        );
     }
 }
